move hparams for cutout and label_smoothing
